Skip impossible tests, nose compatibility, test setup
Some OpenSSL builds can only use some of the ECs, so we should skip testing ecdsa if OpenSSL doesn't support the EC.
Also, nose seems to detect do_test as a test to be run, so remove _test from the helper functions in the OpenSSL test clases.
There is also no need to run openssl version multiple times, so cache the result of the helper function.