VerifyingKey.from_string(): add validate_point= argument
Normally, this defaults to "True", which makes sure that the claimed
verifying key is actually a group element. If you've already validate
this key once (say, before writing it to disk), it is safe to skip this
step, so setting "validate_point=False" will shave a significant amount
of time from the constructor (measured at 26ms per key by @sigmunau).

Closes #19.