🚫🥒 do not pickle the unpickleable RWLock (#199)
* 🚫🥒 do not pickle the unpickleable RWLock

* As requested, acquired the lock before copying PointJacobi. When failing to acquire the lock, it now returns None. Think that's fine or prefer something else?

Added `__ne__` for CurveFp.

I think `test_inaquality_points_diff_types` from test_ellipticcurve.py was not actually testing equality. Because there was no `__ne__`, it was falling back to identity comparison, causing the test to pass. Now that there's a `__ne__`, it was getting NotImplemented which surprisingly bool casts to True. I changed the `__eq__` to return False for different types.

Fix typo in `test_inaquality_points_diff_types` name.

As requested, ran black on modified files. It reformmated some lines I hadn't modified.

* ran black with `--line-length 79` as configured in travis-ci for the project

* Revert to returning NotImplemented.
For __ne__ use == instead of __eq__ so it handles NotImplemented when the types are different.
Remove the not needed state declaration.