move parsing of points from VerifyingKey to parent class of points
For decoding points it's not necessary to have all the data
useful for decoding public keys.

This will also make it possible to decode explicit EC parameters,
as decoding of a public key requires knowledge of the curve's base
point and the base point is in defined in the parameters, creating
a chicken and an egg problem with using the VerifyingKey.from_string()
to parse the base point.